# specs/007-agent-tool-suite/contracts/agent-tools.openapi.yaml
openapi: 3.1.0
info:
  title: Scribe Agent Tools API
  version: 1.0.0
paths:
  /api/github/repo/search:
    get:
      summary: Search text content in the synced repository
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Search results returned
        '400':
          description: Invalid search input or missing assignment
        '404':
          description: Repository checkout unavailable
  /api/github/repo/file:
    put:
      summary: Write a UTF-8 text file inside the synced repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path, content]
              properties:
                path:
                  type: string
                content:
                  type: string
                createDirectories:
                  type: boolean
      responses:
        '200':
          description: File written successfully
        '400':
          description: Invalid path or unsupported content
        '404':
          description: Repository checkout unavailable
  /api/github/repo/git/status:
    get:
      summary: Inspect local git status for the synced repository
      responses:
        '200':
          description: Git status returned
  /api/github/repo/git/diff:
    get:
      summary: Inspect local git diff for the synced repository
      parameters:
        - in: query
          name: path
          schema:
            type: string
      responses:
        '200':
          description: Git diff returned
  /api/github/repo/git/log:
    get:
      summary: Inspect recent git commits for the synced repository
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Git history returned
  /api/github/issues:
    get:
      summary: List open GitHub issues for the assigned repository
      responses:
        '200':
          description: Issue list returned
  /api/github/pulls:
    get:
      summary: List open GitHub pull requests for the assigned repository
      responses:
        '200':
          description: Pull request list returned